edit: file structure
