+++ tests/test_home.py
@@ -1,0 +1,6 @@
+class TestHome:
+    URL = "/"
+
+    def test_get_page(self, client):
+        resp = client.get(self.URL)
+        assert resp.status_code == 200
